+++ plib/controllers/IndexController.php
@@ -1,7 +1,5 @@
 <?php
-/**
- * Copyright 1999-2017. Plesk International GmbH.
- */
+// Copyright 1999-2018. Plesk International GmbH.
 
 /**
  * Class IndexController