+++ plib/library/Helper.php
@@ -1,7 +1,5 @@
 <?php
-/**
- * Copyright 1999-2017. Parallels IP Holdings GmbH.
- */
+// Copyright 1999-2018. Plesk International GmbH.
 
 /**
  * Class Modules_WelcomeBusiness_Helper
@@ -23,7 +21,7 @@ class Modules_WelcomeBusiness_Helper
             $white_list = self::getWhiteListPages();
 
             foreach ($white_list as $item) {
-                if ($url_global_generic.$item == $url_global_return) {
+                if ($url_global_generic . $item == $url_global_return) {
                     return $url_global_return;
                 }
             }
@@ -42,7 +40,7 @@ class Modules_WelcomeBusiness_Helper
         $server_scheme = self::getServerScheme();
         $server_host = self::getServerHost();
 
-        return htmlspecialchars($server_scheme.'://'.$server_host);
+        return htmlspecialchars($server_scheme . '://' . $server_host);
     }
 
     /**
@@ -75,10 +73,10 @@ class Modules_WelcomeBusiness_Helper
         }
 
         if (pm_ProductInfo::isWindows()) {
-            return $_SERVER['LOCAL_ADDR'].':'.$_SERVER['SERVER_PORT'];
+            return $_SERVER['LOCAL_ADDR'] . ':' . $_SERVER['SERVER_PORT'];
         }
 
-        return $_SERVER['SERVER_ADDR'].':'.$_SERVER['SERVER_PORT'];
+        return $_SERVER['SERVER_ADDR'] . ':' . $_SERVER['SERVER_PORT'];
     }
 
     /**
@@ -92,7 +90,7 @@ class Modules_WelcomeBusiness_Helper
             '/admin/',
             '/admin/home?context=home',
             '/smb/',
-            '/smb/web/view'
+            '/smb/web/view',
         );
 
         return $white_list;
@@ -107,7 +105,7 @@ class Modules_WelcomeBusiness_Helper
      */
     public static function isInstalled($name)
     {
-        return file_exists(dirname(pm_Context::getPlibDir()).'/'.$name);
+        return file_exists(dirname(pm_Context::getPlibDir()) . '/' . $name);
     }
 
     /**
@@ -122,7 +120,7 @@ class Modules_WelcomeBusiness_Helper
         $catalog_id = self::getExtensionCatalogId($name);
 
         if (!empty($catalog_id)) {
-            return '/admin/extension/catalog/package/'.$catalog_id;
+            return '/admin/extension/catalog/package/' . $catalog_id;
         }
 
         return '/admin/extension/catalog';
@@ -142,9 +140,10 @@ class Modules_WelcomeBusiness_Helper
             'wp-toolkit'         => '00d002a7-3252-4996-8a08-aa1c89cf29f7',
             'kolab'              => 'bdc94ad6-cc46-4f5a-8ee7-76675fc3cc44',
             'magicspam'          => 'b49f9b1b-e8cf-41e1-bd59-4509d92891f7',
-            'panel-migrator'     => 'bebc4866-d171-45fb-91a6-4b139b8c9a1b',
+            'site-import'        => '01878006-3c3e-4ed6-a7df-37e3741708a2',
             'security-advisor'   => '6bcc01cf-d7bb-4e6a-9db8-dd1826dcad8f',
-            'pagespeed-insights' => '3d2639e6-64a9-43fe-a990-c873b6b3ec66'
+            'pagespeed-insights' => '3d2639e6-64a9-43fe-a990-c873b6b3ec66',
+            'advisor'            => 'bbf16bc7-094e-4cb3-8b9c-32066fc66561',
         );
 
         if (!empty($catalog_ids[$id])) {
@@ -170,7 +169,7 @@ class Modules_WelcomeBusiness_Helper
         $catalog_id = self::getExtensionCatalogId($id);
 
         if (!empty($catalog_id)) {
-            $url = 'https://ext.plesk.com/packages/'.$catalog_id.'-'.$id.'/download';
+            $url = 'https://ext.plesk.com/packages/' . $catalog_id . '-' . $id . '/download';
 
             try {
                 self::installExtensionApi($url);
@@ -205,6 +204,7 @@ class Modules_WelcomeBusiness_Helper
      * Checks whether at least one domain is activated in the subscription
      *
      * @return bool
+     * @throws pm_Exception
      */
     public static function checkAvailableDomains()
     {
@@ -222,6 +222,7 @@ class Modules_WelcomeBusiness_Helper
      * Returns the next step depending on the OS
      *
      * @return int|null|string
+     * @throws pm_Exception
      */
     public static function getNextStep()
     {
@@ -247,6 +248,7 @@ class Modules_WelcomeBusiness_Helper
      * Creates a white list of allowed steps depending on the OS
      *
      * @return array
+     * @throws pm_Exception
      */
     public static function stepListOs()
     {
@@ -254,7 +256,7 @@ class Modules_WelcomeBusiness_Helper
             return array(
                 '1' => 'wp-toolkit',
                 '5' => 'pagespeed-insights',
-                '6' => 'restart'
+                '6' => 'restart',
             );
         }
 
@@ -262,9 +264,9 @@ class Modules_WelcomeBusiness_Helper
             '1' => 'wp-toolkit',
             '2' => 'kolab',
             '3' => 'magicspam',
-            '4' => 'security-advisor',
+            '4' => self::getAdvisorData(),
             '5' => 'pagespeed-insights',
-            '6' => 'restart'
+            '6' => 'restart',
         );
     }
 
@@ -282,6 +284,55 @@ class Modules_WelcomeBusiness_Helper
         }
 
         if (time() - $executed >= 3) {
+            return true;
+        }
+
+        return false;
+    }
+
+    /**
+     * Gets the correct link to the new domain creation page
+     *
+     * @return string
+     * @throws pm_Exception
+     */
+    public static function getAdvisorData($element = 'id')
+    {
+        $advisorExtension = ['name' => 'Security Advisor', 'id' => 'security-advisor'];
+
+        if (self::isPleskVersion178()) {
+            $advisorExtension = ['name' => 'Advisor', 'id' => 'advisor'];
+        }
+
+        return $advisorExtension[$element];
+    }
+
+    /**
+     * Gets the correct link to the new domain creation page
+     *
+     * @return string
+     * @throws pm_Exception
+     */
+    public static function getLinkNewDomain()
+    {
+        if (self::isPleskVersion178()) {
+            return '/smb/web/add-domain';
+        }
+
+        return '/admin/subscription/create';
+    }
+
+    /**
+     * Checks whether Plesk version is >= 17.8.10
+     *
+     * @return bool
+     * @throws pm_Exception
+     */
+    private static function isPleskVersion178()
+    {
+        $pleskVersion = pm_ProductInfo::getVersion();
+
+        if (version_compare($pleskVersion, '17.8.10', 'ge')) {
             return true;
         }