+++ CHANGES.md
@@ -1,3 +1,9 @@
-# 1.0.0 (25 March 2017)
+# 1.0.1
+
+* [+] Compatibility with Plesk Onyx 17.8
+* [*] Replaced "Security Advisor" with "Advisor" in Plesk Onyx 17.8
+* [*] Replaced "Site Migrator" with "Site Import"
+
+# 1.0.0
 
 * [+] First release of the "Welcome - Business Server" extension